Dataset_Construction/requirements.txt
langchain==0.0.183
openai==0.27.7
pinecone_client==2.2.1
PyYAML==6.0
tqdm==4.65.0
